Visual improvements
The following visual improvements have been implemented:
A alternate environment will display when playing endless.
Pause and end-of-run messages now display in a "window."
Buttons to restart and return to main are available when either of these windows are active.
The mouse cursor is now hidden and disabled when appropriate.
HUD font styles have been adjusted.
Particle effects have been added for the player's slide and dive moves.
Slight adjustments to animation transition times.

Aside the visual stuff, the Saw obstacle now plays audio as it travels across the screen.